Enhance AppSettings validation with data annotations
Updated the `AddAppSettingsFrom` method in `AppSettingsGenerator.cs` to support data annotation validation for options. If no validator class is provided, it now configures options with validation on start. If a validator class is specified, a scoped service for the validator is added along with FluentValidation configuration. Also, added the `Microsoft.Extensions.Options.DataAnnotations` package to the project for data annotation support.